Add files via upload
This was a guided python project provided by dataquest. I believe it represented a substantial change in the confidence i have in coding. For the first time i was able to progress through a project with almost no use of the solutions handbook, independently debugging issues i faced myself and even utilising the community forum to help others with issues they had which were similar to mine. Onwards and upwards from here, God willing.